test: split tests for scenario-management template to a separate file
